+++ apps/prismatic_runtime/lib/prismatic/adapters/oauth_browser/system_cmd.ex
@@ -1,0 +1,10 @@
+defmodule Prismatic.Adapters.OAuthBrowser.SystemCmd do
+  @moduledoc """
+  System-command browser launcher adapter for interactive OAuth flows.
+  """
+
+  @behaviour Prismatic.Ports.OAuthBrowser
+
+  @impl true
+  defdelegate open(url, opts \\ []), to: Prismatic.OAuth2.Browser
+end